refactor: refactor score calcultor inside Hand model
