+++ src/main/kotlin/model/Hand.kt
@@ -1,7 +1,5 @@
 package model
 
-import service.ScoreCalculator
-
 class Hand {
     private val cards: MutableSet<Card> = mutableSetOf()
 
@@ -12,6 +10,24 @@ class Hand {
     }
 
     fun scoreOnHand(): Int {
-        return ScoreCalculator.getScore(cards)
+        return getScore(cards)
+    }
+
+    companion object {
+        fun getScore(cards: Set<Card>): Int {
+            var scoreTotal = scoreSum(cards)
+            var aceCounter = aceCount(cards)
+            while (scoreTotal > 21 && hasAce(aceCounter)) {
+                scoreTotal -= 10
+                aceCounter--
+            }
+            return scoreTotal
+        }
+
+        private fun hasAce(aceCounter: Int): Boolean = aceCounter != 0
+
+        private fun scoreSum(cards: Set<Card>) = cards.sumOf { card -> Rank.score(card.rank) }
+
+        private fun aceCount(cards: Set<Card>) = cards.count { it.rank == Rank.ACE }
     }
 }
