+++ src/test/kotlin/model/HandTest.kt
@@ -3,6 +3,7 @@ package model
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 
 class HandTest {
@@ -15,40 +16,68 @@ class HandTest {
     }
 
     @ParameterizedTest
+    @MethodSource("cardProvider")
+    fun `score for one card should return card value`(card: Card) {
+        assertThat(Hand.getScore(setOf(card))).isEqualTo(Rank.score(card.rank))
+    }
+
+    @ParameterizedTest
     @MethodSource("setCardProvider")
-    fun `test score on hand`(cards: Set<Card>) {
-        val hand = Hand()
-        cards.forEach { hand.addCard(it) }
-        val previousScore = hand.scoreOnHand()
-        val card = Card(Rank.THREE, Suite.HEARTS)
-        hand.addCard(card)
-        assertThat(previousScore + Rank.score(card.rank)).isEqualTo(hand.scoreOnHand())
+    fun `score for set of cards should return total value of scores`(
+        cards: Set<Card>,
+        expectedTotal: Int,
+    ) {
+        assertThat(Hand.getScore(cards)).isEqualTo(expectedTotal)
     }
 
     companion object {
         @JvmStatic
-        fun setCardProvider(): List<Set<Card>> {
+        fun cardProvider() = listOf(Card(Rank.EIGHT, Suite.DIAMONDS), Card(Rank.QUEEN, Suite.DIAMONDS), Card(Rank.TWO, Suite.DIAMONDS))
+
+        @JvmStatic
+        fun setCardProvider(): List<Arguments> {
             return listOf(
-                setOf(
-                    Card(Rank.THREE, Suite.DIAMONDS),
-                    Card(Rank.NINE, Suite.CLUBS),
-                    Card(Rank.TWO, Suite.DIAMONDS),
+                Arguments.of(
+                    setOf(
+                        Card(Rank.THREE, Suite.DIAMONDS),
+                        Card(Rank.NINE, Suite.CLUBS),
+                        Card(Rank.EIGHT, Suite.DIAMONDS),
+                    ),
+                    20,
                 ),
-                setOf(
-                    Card(Rank.TWO, Suite.HEARTS),
-                    Card(Rank.EIGHT, Suite.SPADES),
-                    Card(Rank.FIVE, Suite.CLUBS),
+                Arguments.of(
+                    setOf(
+                        Card(Rank.TWO, Suite.HEARTS),
+                        Card(Rank.EIGHT, Suite.SPADES),
+                        Card(Rank.ACE, Suite.CLUBS),
+                    ),
+                    21,
                 ),
-                setOf(
-                    Card(Rank.TWO, Suite.DIAMONDS),
-                    Card(Rank.THREE, Suite.CLUBS),
-                    Card(Rank.FOUR, Suite.HEARTS),
+                Arguments.of(
+                    setOf(
+                        Card(Rank.ACE, Suite.DIAMONDS),
+                        Card(Rank.ACE, Suite.CLUBS),
+                        Card(Rank.ACE, Suite.HEARTS),
+                    ),
+                    13,
                 ),
-                setOf(
-                    Card(Rank.FIVE, Suite.CLUBS),
-                    Card(Rank.SIX, Suite.HEARTS),
-                    Card(Rank.TWO, Suite.SPADES),
-                    Card(Rank.TWO, Suite.CLUBS),
+                Arguments.of(
+                    setOf(
+                        Card(Rank.ACE, Suite.CLUBS),
+                        Card(Rank.ACE, Suite.HEARTS),
+                        Card(Rank.ACE, Suite.SPADES),
+                        Card(Rank.EIGHT, Suite.CLUBS),
+                    ),
+                    21,
+                ),
+                Arguments.of(
+                    setOf(
+                        Card(Rank.ACE, Suite.CLUBS),
+                        Card(Rank.ACE, Suite.HEARTS),
+                        Card(Rank.TEN, Suite.SPADES),
+                        Card(Rank.TEN, Suite.CLUBS),
+                    ),
+                    22,
                 ),
             )
         }
